requirements.txt
# requirements.txt

mcp>=1.9.4            # FastMCP + the full MCP SDK
uvicorn>=0.27.0       # ASGI server
slack-sdk>=3.21.0     # Slack WebClient
python-dotenv>=1.0.0  # loads .env locally
flask>=3.0.0          # OAuth web server
fastapi>=0.109.0      # FastAPI for HTTP transport
